+++ src/Controller/User/Base.php
@@ -8,6 +8,8 @@ use App\Controller\BaseController;
 use App\Exception\User;
 use App\Service\User\Find;
 use App\Service\User\Create;
+use App\Service\User\Update;
+use App\Service\User\Delete;
 use App\Service\User\UserService;
 
 abstract class Base extends BaseController
@@ -25,6 +27,16 @@ abstract class Base extends BaseController
     protected function getCreateUserService(): Create
     {
         return $this->container->get('create_user_service');
+    }
+
+    protected function getUpdateUserService(): Update
+    {
+        return $this->container->get('update_user_service');
+    }
+
+    protected function getDeleteUserService(): Delete
+    {
+        return $this->container->get('delete_user_service');
     }
 
     protected function checkUserPermissions(int $userId, int $userIdLogged): void
