+++ src/Controller/User/Update.php
@@ -14,7 +14,7 @@ final class Update extends Base
         $input = (array) $request->getParsedBody();
         $userIdLogged = $this->getAndValidateUserId($input);
         $this->checkUserPermissions((int) $args['id'], (int) $userIdLogged);
-        $user = $this->getUserService()->update($input, (int) $args['id']);
+        $user = $this->getUpdateUserService()->update($input, (int) $args['id']);
 
         return $this->jsonResponse($response, 'success', $user, 200);
     }
